Make InstallationTarget objects self contained
Closes #5. You can read all of the context there too.

We want InstallationTarget objects to be self-contained, independent
from releases and applications. We're making them so by applying a few
changes:

1. The installation controller loses its smarts in regards to finding
out if the InstallationTarget is part of an incumbent or contender
release. This logic now belongs entiretly to the release controller.

2. The InstallationTarget becomes a richer object, by containing the
chart and values specified by the Application object at the time of the
creation of the release.

This commit takes a slightly different approach than the one described
in #5, though. Since we currently don't support installing different
things (either different charts or different values) per cluster with
the same InstallationTarget, we chose not to implement it here. That's
mostly in case YAGNI, and the cost of doing so in the future is roughly
the same as doing it now, so we took the call to punt it to future us.

We also decided to continue completely skipping InstallationTargets that
do not belong to a contender release (or that have CanOverride = false,
now), because that would slow down the installation controller too much
before #77 is closed. We captured that requirement in #110 though.