Replaced ghodss/yaml with community-supported sigs.k8s.io/yaml
This commit is a followup on a conversation in
https://github.com/ghodss/yaml/issues/50. In essence, the change is
about switching to the community-supported clone of yaml library.
This change is quite wanted as it moves Shipper forward to no dependency
on the aforementioned non-supported library. At present, there are quite
some dependencies left which are mainly required by outdated versions of
supporting libraries shipper uses. Once all the referred projects
transition to the newer version of the library, Shipper can finalize the
upgrade.

Signed-off-by: Oleg Sidorov <oleg.sidorov@booking.com>